refactor(room): remove ngrx-effects for StartGame() action
