+++ src/app/game/game-over-view/game-over-view.component.html
@@ -21,7 +21,7 @@
                  class="u-marginBottom-20 alt-theme">
                 <button mat-raised-button
                         color="primary"
-                        (click)="playerAgain()"
+                        (click)="playAgain(room)"
                         class="u-width-100 c-btn c-btn-bottom-border--light-blue">
                     <div [innerHTML]="'Play Again'"></div>
                 </button>
