+++ src/app/room/room.reducer.ts
@@ -1,8 +1,6 @@
 import { Room } from '../interfaces/room.model';
 import {
     RoomActions,
-    START_GAME_SUCCESS_TYPE,
-    START_GAME_TYPE,
     UPDATE_ROOM,
     UPDATE_ROOM_SUCCESS,
 } from './room.actions';
@@ -21,19 +19,6 @@ export function roomReducer(state: Room, action: RoomActions) {
                 ...state,
                 ...action.payload,
                 loading: false,
-            };
-
-        case START_GAME_TYPE:
-            return {
-                ...state,
-                loading: true,
-            };
-
-        case START_GAME_SUCCESS_TYPE:
-            return {
-                ...state,
-                loading: false,
-                ...action.payload,
             };
 
         default:
